apps/ui: split onboarding route into per-page folder
Breaks the onboarding.tsx monolith (layout shell + home + create + css)
into:

  router/onboarding/
    index.tsx        — route tree + OnboardingShell
    home.tsx         — two-card landing page
    create.tsx       — create-site form page
    style.module.css — shared page styles

Each page file exports just the component; index.tsx owns the
createRoute wiring. Other routes (dashboard, sessions) stay as single
files since they're small enough.
