+++ apps/ui/src/router/onboarding/create.tsx
@@ -1,0 +1,59 @@
+import { useNavigate } from '@tanstack/react-router';
+import { __ } from '@wordpress/i18n';
+import { useState } from 'react';
+import { CreateSiteForm } from '@/components/create-site-form';
+import {
+	useExistingCustomDomains,
+	useProposedSiteName,
+} from '@/data/queries/use-create-site-helpers';
+import { useCreateSite, useSites } from '@/data/queries/use-sites';
+import styles from './style.module.css';
+import type { CreateSiteFormValues } from '@/components/create-site-form';
+
+export function CreateSitePage() {
+	const navigate = useNavigate();
+	const { data: sites } = useSites();
+	const { data: existingDomainNames } = useExistingCustomDomains();
+	const { data: proposedName } = useProposedSiteName( sites );
+	const createSite = useCreateSite();
+	const [ submitError, setSubmitError ] = useState( '' );
+
+	const handleSubmit = async ( values: CreateSiteFormValues ) => {
+		setSubmitError( '' );
+		try {
+			await createSite.mutateAsync( {
+				name: values.name,
+				path: values.path,
+				phpVersion: values.phpVersion,
+				wpVersion: values.wpVersion,
+				customDomain: values.customDomain,
+				enableHttps: values.enableHttps,
+				adminUsername: values.adminUsername || undefined,
+				adminPassword: values.adminPassword || undefined,
+				adminEmail: values.adminEmail || undefined,
+			} );
+			await navigate( { to: '/dashboard' } );
+		} catch ( error ) {
+			setSubmitError(
+				error instanceof Error ? error.message : __( 'Failed to create site. Please try again.' )
+			);
+		}
+	};
+
+	return (
+		<div className={ styles.page }>
+			<h1 className={ styles.title }>{ __( 'Create a new site' ) }</h1>
+			<p className={ styles.subtitle }>
+				{ __( 'Choose a name and we\u2019ll scaffold a fresh WordPress site locally.' ) }
+			</p>
+			<CreateSiteForm
+				defaultName={ proposedName }
+				existingDomainNames={ existingDomainNames ?? [] }
+				onSubmit={ handleSubmit }
+				onCancel={ () => void navigate( { to: '/onboarding' } ) }
+				isSubmitting={ createSite.isPending }
+				submitError={ submitError }
+			/>
+		</div>
+	);
+}
